Update all the things
This repo was shared a bit prematurely. What previously existed was
essentially the React tutorial with a bit of WordPress flavouring.

This update contains a lot of the basic functionality that can now be
used to build a theme.

The inline documentation is non-existent but I am going to be going
through and sorting these things out as I get to them.

There are a lot of to-dos and bits that are nowhere near ready but I
thought it would be good to get something pushed!